conffile: add system wide configuration file
In some cases, where the configuration file is not tied to a user, it
can be useful to have system wide configuration.
This enables the tool to look for conffiles in /etc/pw/ if nothing is
found in ${HOME}.
Note that, with this addition, configuration files can now be shipped
along with the tool in packages (.rpm, .deb, ...).

Signed-off-by: Liam Beguin <liambeguin@gmail.com>