pw: conform to pep8 standard
This is a major code cleanup. It makes the code pep8 compliant but is a
major pain from a Git standpoint since the switch from tabs to spaces
results in a +80% rewrite of the file.

While at it, add a little documentation to various functions and update
the usage message to include [options]

Signed-off-by: Liam Beguin <liambeguin@gmail.com>