pw: extract version from git
instead of having to update the tool's revision by hand, lets extract it
using `git describe`. In the event that the tool is no longer in a git
repository, the version function will output %%VERSION%%. This is a unique
string that can be replace by a futur `make install` command.

Signed-off-by: Liam Beguin <liambeguin@gmail.com>