conffile: allow use of system configuration files
when the tool is installed using a package manager, it can be useful to
ship configuration files along with it. This allows the tool to search
for configuration files in /etc/pw/ and in $HOME.

As a reminder, the tool looks for a configuration file based on $0. This
means that using symlinks allows the user to control multiple switches
with ease.

Signed-off-by: Liam Beguin <liambeguin@gmail.com>